Refactor card backs and tests
