+++ Assets/Tests/PlayMode/CardGameManagerTests.cs
@@ -7,7 +7,6 @@ using Cgs;
 using FinolDigital.Cgs.Json;
 using FinolDigital.Cgs.Json.Unity;
 using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 using Newtonsoft.Json.Serialization;
 using NUnit.Framework;
 using UnityEngine;
@@ -169,101 +168,5 @@ namespace Tests.PlayMode
             Assert.AreEqual("Dominoes", CardGameManager.Current.Name);
         }
 
-        [Test]
-        public void WriteAllCardsJson_WritesSchemaAlignedBacks()
-        {
-            var game = new UnityCardGame(null, "write_backs_test_" + Guid.NewGuid())
-            {
-                CardProperties = new List<PropertyDef>()
-            };
-
-            if (Directory.Exists(game.GameDirectoryPath))
-                Directory.Delete(game.GameDirectoryPath, true);
-            Directory.CreateDirectory(game.GameDirectoryPath);
-
-            var withBack = new UnityCard(game, "card_with_back", "With Back", Set.DefaultCode,
-                new Dictionary<string, PropertyDefValuePair>(), false, false, "DARK_BACK");
-            var defaultBack = new UnityCard(game, "card_default_back", "Default Back", Set.DefaultCode,
-                new Dictionary<string, PropertyDefValuePair>(), false);
-
-            game.Add(withBack, false);
-            game.Add(defaultBack, false);
-            game.WriteAllCardsJson();
-
-            var allCards = JArray.Parse(File.ReadAllText(game.CardsFilePath));
-            var withBackToken = allCards.Children<JObject>().First(token =>
-                string.Equals(token.Value<string>("id"), "card_with_back", StringComparison.Ordinal));
-            var defaultBackToken = allCards.Children<JObject>().First(token =>
-                string.Equals(token.Value<string>("id"), "card_default_back", StringComparison.Ordinal));
-
-            Assert.AreEqual("DARK_BACK", withBackToken["backs"]?[0]?.Value<string>());
-            Assert.IsNull(withBackToken["backFaceId"]);
-            Assert.IsNull(defaultBackToken["backs"]);
-            Assert.IsNull(defaultBackToken["backFaceId"]);
-
-            Directory.Delete(game.GameDirectoryPath, true);
-        }
-
-        [Test]
-        public void LoadCards_UsesBacksThenBackFaceIdFallback()
-        {
-            var game = new UnityCardGame(null, "load_backs_test_" + Guid.NewGuid())
-            {
-                CardProperties = new List<PropertyDef>()
-            };
-
-            if (Directory.Exists(game.GameDirectoryPath))
-                Directory.Delete(game.GameDirectoryPath, true);
-            Directory.CreateDirectory(game.GameDirectoryPath);
-
-            var allCards = new JArray
-            {
-                new JObject
-                {
-                    ["id"] = "legacy_back",
-                    ["name"] = "Legacy",
-                    ["set"] = Set.DefaultCode,
-                    ["backFaceId"] = "LEGACY_BACK"
-                },
-                new JObject
-                {
-                    ["id"] = "canonical_back",
-                    ["name"] = "Canonical",
-                    ["set"] = Set.DefaultCode,
-                    ["backs"] = new JArray("CANON_BACK")
-                },
-                new JObject
-                {
-                    ["id"] = "both_back",
-                    ["name"] = "Both",
-                    ["set"] = Set.DefaultCode,
-                    ["backs"] = new JArray("PREFERRED_BACK"),
-                    ["backFaceId"] = "LEGACY_SHOULD_NOT_WIN"
-                },
-                new JObject
-                {
-                    ["id"] = "default_back",
-                    ["name"] = "Default",
-                    ["set"] = Set.DefaultCode
-                }
-            };
-
-            File.WriteAllText(game.CardsFilePath, allCards.ToString(Formatting.None));
-            game.LoadCards(game.CardsFilePath, Set.DefaultCode);
-
-            Assert.IsTrue(game.Cards.TryGetValue("legacy_back", out var legacyCard));
-            Assert.AreEqual("LEGACY_BACK", legacyCard.BackFaceId);
-
-            Assert.IsTrue(game.Cards.TryGetValue("canonical_back", out var canonicalCard));
-            Assert.AreEqual("CANON_BACK", canonicalCard.BackFaceId);
-
-            Assert.IsTrue(game.Cards.TryGetValue("both_back", out var bothCard));
-            Assert.AreEqual("PREFERRED_BACK", bothCard.BackFaceId);
-
-            Assert.IsTrue(game.Cards.TryGetValue("default_back", out var defaultCard));
-            Assert.IsTrue(string.IsNullOrEmpty(defaultCard.BackFaceId));
-
-            Directory.Delete(game.GameDirectoryPath, true);
-        }
     }
 }
